+++ transactron/lib/methods.py
@@ -1,0 +1,267 @@
+from amaranth import *
+from ..core import *
+from ..core import RecordDict
+from typing import Optional, Callable, Tuple
+from coreblocks.utils import ValueLike, assign, AssignType
+from .transactions import ManyToOneConnectTrans
+from .connections import Forwarder
+
+__all__ = [
+    "MethodTransformer",
+    "MethodFilter",
+    "MethodProduct",
+    "MethodTryProduct",
+    "Collector",
+    ]
+
+class MethodTransformer(Elaboratable):
+    """Method transformer.
+
+    Takes a target method and creates a transformed method which calls the
+    original target method, transforming the input and output values.
+    The transformation functions take two parameters, a `Module` and the
+    `Record` being transformed. Alternatively, a `Method` can be
+    passed.
+
+    Attributes
+    ----------
+    method: Method
+        The transformed method.
+    """
+
+    def __init__(
+        self,
+        target: Method,
+        *,
+        i_transform: Optional[Tuple[MethodLayout, Callable[[TModule, Record], RecordDict]]] = None,
+        o_transform: Optional[Tuple[MethodLayout, Callable[[TModule, Record], RecordDict]]] = None,
+    ):
+        """
+        Parameters
+        ----------
+        target: Method
+            The target method.
+        i_transform: (record layout, function or Method), optional
+            Input transformation. If specified, it should be a pair of a
+            function and a input layout for the transformed method.
+            If not present, input is not transformed.
+        o_transform: (record layout, function or Method), optional
+            Output transformation. If specified, it should be a pair of a
+            function and a output layout for the transformed method.
+            If not present, output is not transformed.
+        """
+        if i_transform is None:
+            i_transform = (target.data_in.layout, lambda _, x: x)
+        if o_transform is None:
+            o_transform = (target.data_out.layout, lambda _, x: x)
+
+        self.target = target
+        self.method = Method(i=i_transform[0], o=o_transform[0])
+        self.i_fun = i_transform[1]
+        self.o_fun = o_transform[1]
+
+    def elaborate(self, platform):
+        m = TModule()
+
+        @def_method(m, self.method)
+        def _(arg):
+            return self.o_fun(m, self.target(m, self.i_fun(m, arg)))
+
+        return m
+
+
+class MethodFilter(Elaboratable):
+    """Method filter.
+
+    Takes a target method and creates a method which calls the target method
+    only when some condition is true. The condition function takes two
+    parameters, a module and the input `Record` of the method. Non-zero
+    return value is interpreted as true. Alternatively to using a function,
+    a `Method` can be passed as a condition.
+
+    Caveat: because of the limitations of transaction scheduling, the target
+    method is locked for usage even if it is not called.
+
+    Attributes
+    ----------
+    method: Method
+        The transformed method.
+    """
+
+    def __init__(
+        self, target: Method, condition: Callable[[TModule, Record], ValueLike], default: Optional[RecordDict] = None
+    ):
+        """
+        Parameters
+        ----------
+        target: Method
+            The target method.
+        condition: function or Method
+            The condition which, when true, allows the call to `target`. When
+            false, `default` is returned.
+        default: Value or dict, optional
+            The default value returned from the filtered method when the condition
+            is false. If omitted, zero is returned.
+        """
+        if default is None:
+            default = Record.like(target.data_out)
+
+        self.target = target
+        self.method = Method.like(target)
+        self.condition = condition
+        self.default = default
+
+    def elaborate(self, platform):
+        m = TModule()
+
+        ret = Record.like(self.target.data_out)
+        m.d.comb += assign(ret, self.default, fields=AssignType.ALL)
+
+        @def_method(m, self.method)
+        def _(arg):
+            with m.If(self.condition(m, arg)):
+                m.d.comb += ret.eq(self.target(m, arg))
+            return ret
+
+        return m
+
+
+class MethodProduct(Elaboratable):
+    def __init__(
+        self,
+        targets: list[Method],
+        combiner: Optional[Tuple[MethodLayout, Callable[[TModule, list[Record]], RecordDict]]] = None,
+    ):
+        """Method product.
+
+        Takes arbitrary, non-zero number of target methods, and constructs
+        a method which calls all of the target methods using the same
+        argument. The return value of the resulting method is, by default,
+        the return value of the first of the target methods. A combiner
+        function can be passed, which can compute the return value from
+        the results of every target method.
+
+        Parameters
+        ----------
+        targets: list[Method]
+            A list of methods to be called.
+        combiner: (int or method layout, function), optional
+            A pair of the output layout and the combiner function. The
+            combiner function takes two parameters: a `Module` and
+            a list of outputs of the target methods.
+
+        Attributes
+        ----------
+        method: Method
+            The product method.
+        """
+        if combiner is None:
+            combiner = (targets[0].data_out.layout, lambda _, x: x[0])
+        self.targets = targets
+        self.combiner = combiner
+        self.method = Method(i=targets[0].data_in.layout, o=combiner[0])
+
+    def elaborate(self, platform):
+        m = TModule()
+
+        @def_method(m, self.method)
+        def _(arg):
+            results = []
+            for target in self.targets:
+                results.append(target(m, arg))
+            return self.combiner[1](m, results)
+
+        return m
+
+
+class MethodTryProduct(Elaboratable):
+    def __init__(
+        self,
+        targets: list[Method],
+        combiner: Optional[tuple[MethodLayout, Callable[[TModule, list[tuple[Value, Record]]], RecordDict]]] = None,
+    ):
+        """Method product with optional calling.
+
+        Takes arbitrary, non-zero number of target methods, and constructs
+        a method which tries to call all of the target methods using the same
+        argument. The methods which are not ready are not called. The return
+        value of the resulting method is, by default, empty. A combiner
+        function can be passed, which can compute the return value from the
+        results of every target method.
+
+        Parameters
+        ----------
+        targets: list[Method]
+            A list of methods to be called.
+        combiner: (int or method layout, function), optional
+            A pair of the output layout and the combiner function. The
+            combiner function takes two parameters: a `Module` and
+            a list of pairs. Each pair contains a bit which signals
+            that a given call succeeded, and the result of the call.
+
+        Attributes
+        ----------
+        method: Method
+            The product method.
+        """
+        if combiner is None:
+            combiner = ([], lambda _, __: {})
+        self.targets = targets
+        self.combiner = combiner
+        self.method = Method(i=targets[0].data_in.layout, o=combiner[0])
+
+    def elaborate(self, platform):
+        m = TModule()
+
+        @def_method(m, self.method)
+        def _(arg):
+            results: list[tuple[Value, Record]] = []
+            for target in self.targets:
+                success = Signal()
+                with Transaction().body(m):
+                    m.d.comb += success.eq(1)
+                    results.append((success, target(m, arg)))
+            return self.combiner[1](m, results)
+
+        return m
+
+class Collector(Elaboratable):
+    """Single result collector.
+
+    Creates method that collects results of many methods with identical
+    layouts. Each call of this method will return a single result of one
+    of the provided methods.
+
+    Attributes
+    ----------
+    method: Method
+        Method which returns single result of provided methods.
+    """
+
+    def __init__(self, targets: list[Method]):
+        """
+        Parameters
+        ----------
+        method_list: list[Method]
+            List of methods from which results will be collected.
+        """
+        self.method_list = targets
+        layout = targets[0].data_out.layout
+        self.method = Method(o=layout)
+
+        for method in targets:
+            if layout != method.data_out.layout:
+                raise Exception("Not all methods have this same layout")
+
+    def elaborate(self, platform):
+        m = TModule()
+
+        m.submodules.forwarder = forwarder = Forwarder(self.method.data_out.layout)
+
+        m.submodules.connect = ManyToOneConnectTrans(
+            get_results=[get for get in self.method_list], put_result=forwarder.write
+        )
+
+        self.method.proxy(m, forwarder.read)
+
+        return m
